+++ src/zenboard/utils.py
@@ -1,10 +1,15 @@
 """
 zenboard utils
 """
+import logging
+
 from django.conf import settings
 from github3 import GitHub
 
 from zenhub_api import ZenHubAPI
+
+
+logger = logging.getLogger(__name__)
 
 
 class ValidateModelMixin:
@@ -18,9 +23,16 @@ class ValidateModelMixin:
         return attrs
 
 
+if not settings.GITHUB_TOKEN:
+    logging.warning("GitHub API token not found")
+
 github_api = GitHub(
     token=settings.GITHUB_TOKEN,
 )
+
+
+if not settings.GITHUB_TOKEN:
+    logging.warning("ZenHub API token not found")
 
 zenhub_api = ZenHubAPI(
     token=settings.ZENHUB_API_TOKEN,