+++ src/zenboard/__init__.py
@@ -3,7 +3,7 @@ Zenboard - Read only ZenHub boards made easy
 """
 
 __title__ = 'Zenboard'
-__version__ = '0.1.2.dev0'
+__version__ = '0.2.0'
 __author__ = 'Paweł Adamczak'
 __license__ = 'Apache 2'
 __url__ = 'https://github.com/pawelad/zenboard'