+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][keepachangelog] and this project
 adheres to [Semantic Versioning][semver].
 
 ## [Unreleased][unreleased]
+
+## [v0.2.0][v0.2.0] - 2017-10-31
 ### Added
 - Renamed `Board.name` field to `slug` and added proper `name` field.
 - Added `Board.show_closed_pipeline` field.
@@ -36,9 +38,10 @@ adheres to [Semantic Versioning][semver].
 [keepachangelog]: http://keepachangelog.com/en/1.0.0/
 [semver]: http://semver.org/spec/v2.0.0.html
 
-[unreleased]: https://github.com/pawelad/zenboard/compare/v0.1.1...HEAD
+[unreleased]: https://github.com/pawelad/zenboard/compare/v0.2.0...HEAD
 [v0.1.0]: https://github.com/pawelad/zenboard/releases/tag/v0.1.0
-[v0.1.1]: https://github.com/pawelad/zenboard/releases/tag/v0.1.0
+[v0.1.1]: https://github.com/pawelad/zenboard/releases/tag/v0.1.1
+[v0.2.0]: https://github.com/pawelad/zenboard/releases/tag/v0.2.0
 
 [#1]: https://github.com/pawelad/zenboard/issues/1
 [#2]: https://github.com/pawelad/zenboard/issues/2